+++ bin/datasafe_help.sh
@@ -5,7 +5,7 @@
 # Script.....: datasafe_help.sh
 # Author.....: Stefan Oehrli (oes) stefan.oehrli@oradba.ch
 # Date.......: 2026.02.16
-# Version....: v0.9.2
+# Version....: v0.11.2
 # Purpose....: Wrapper for odb_datasafe_help.sh
 # License....: Apache License Version 2.0
 # ------------------------------------------------------------------------------